Gather ui adapter code into fewer files
