+++ src/ai/agents/ui/ai_sdk/__init__.py
@@ -1,12 +1,13 @@
 """AI SDK UI adapter for messages and SSE streams."""
 
-from .inbound import (
+from .approvals import (
     ApprovalResponse,
     apply_approvals,
     extract_approvals,
-    to_messages,
 )
-from .outbound import to_sse, to_stream, to_ui_messages
+from .inbound_messages import to_messages
+from .outbound_messages import to_ui_messages
+from .outbound_stream import to_sse, to_stream
 from .ui_events import UI_MESSAGE_STREAM_HEADERS
 from .ui_messages import UIMessage
